fix: 修复双总结（过程回复未标记 React 阶段）+ 计时移至执行状态左侧
问题分析（结合 session 03gdu1w6 的数据）：
一次轮次中，LLM 在工具执行阶段输出文本回复后未调用工具，该过程文本被持久化
但未标记 phase=React（保持 normal）；随后总结阶段又生成一份最终回复。两条
assistant 回复均为 phase=normal，前端将两者都作为主回复渲染，导致「先显示详实
总结、完成后被简短总结替换」的现象。

修复：
- engine.rs：ReAct 阶段无工具调用的过程回复标记 phase=React，前端据此将其折叠
  为过程内容，仅展示总结阶段的最终回复
- 将推理实时计时从 MessageList（流式/思考块）移至 MessageInput「执行状态」
  指示器（Circle + runSummary）左侧，与状态文案同组展示
